+++ apps/web/src/components/scenario-form/scenario-form.tsx
@@ -23,12 +23,6 @@ export function ScenarioForm() {
       aid: getRandomCallsign(),
       cid: getRandomCid(),
       bcn: getRandomBcn(),
-      typ: "B739",
-      eq: "L",
-      dep: "KPDX",
-      dest: "KLAS",
-      spd: 225,
-      alt: "350",
       vatsimId: getRandomVatsimId(),
       pilotName: getRandomName(),
     },
